Squashed commit of the following:

commit fae5565f41fb7db83287999210d7dda6457b5685
Author: Russ Combs <rucombs@cisco.com>
Date:   Fri Dec 9 16:27:32 2016 -0500

    fix cppcheck warning in ftp params

commit f41129dd4622688af4899c1faf9e45b33d90162a
Author: Russ Combs <rucombs@cisco.com>
Date:   Fri Dec 9 15:12:15 2016 -0500

    convert stream u2 maps arg to reference

commit 3347756cf56413da5a5fdd4210918bb7257a2f86
Author: Russ Combs <rucombs@cisco.com>
Date:   Fri Dec 9 12:53:28 2016 -0500

    fix int vs size_t format issue

commit 01ab793e28c8971e2d1094a4b1ea44bf7f68d943
Author: Russ Combs <rucombs@cisco.com>
Date:   Fri Dec 9 12:34:51 2016 -0500

    remove vestigial extra stream debug foo

+++ src/actions/act_react.cc
@@ -230,7 +230,7 @@ static bool react_getpage(const char* file)
 // format response buffer
 static void react_config(ReactData* rd)
 {
-    size_t body_len, head_len, total_len;
+    int body_len, head_len, total_len;
     char dummy;
 
     const char* head = DEFAULT_HTTP;
